requirements.txt
# Core dependencies
fastapi==0.116.1
uvicorn[standard]==0.24.0
pydantic==2.8.0
python-dotenv==1.0.0

# LLM providers
openai==1.95.1
anthropic==0.57.1

# Data processing
pandas==2.1.3
numpy==1.26.0
yfinance==0.2.33

# MCP and WebSocket
mcp==1.11.0
websockets==12.0
aiohttp==3.9.0

# Redis for context storage
redis==5.0.1

# Utilities
pyyaml==6.0.1
python-dateutil==2.8.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
anyio==4.5.0